Renamed function to change point tests. Closes #1391.
