+++ pints/cptests/_problems.py
@@ -1,5 +1,5 @@
 #
-# Shared problems used in functional testing.
+# Shared problems used in change point testing.
 #
 # This file is part of PINTS (https://github.com/pints-team/pints/) which is
 # released under the BSD 3-clause license. See accompanying LICENSE.md for
