+++ pints/cptests/dram_acmc.py
@@ -1,88 +1,69 @@
 #!/usr/bin/env python3
 #
-# Functional tests for DreamMCMC
+# Change point tests for DramACMC
 #
 # This file is part of PINTS (https://github.com/pints-team/pints/) which is
 # released under the BSD 3-clause license. See accompanying LICENSE.md for
 # copyright notice and full license details.
 #
 import pints
-import pints.functionaltests as ft
+import pints.cptests as ft
 
 
-def two_dim_gaussian(n_iterations=10000, n_warmup=1000):
+def two_dim_gaussian(n_iterations=8000, n_warmup=2000):
     """
-    Tests :class:`pints.DreamMCMC`
+    Tests :class:`pints.DramACMC`
     on a two-dimensional Gaussian distribution with true solution
     ``[0, 0]`` and returns a dictionary with entries ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnTwoDimGaussian`.
+    :class:`pints.cptests.RunMcmcMethodOnTwoDimGaussian`.
     """
     problem = ft.RunMcmcMethodOnTwoDimGaussian(
-        _method, 10, n_iterations, n_warmup)
-
+        _method, 4, n_iterations, n_warmup)
     return {
         'kld': problem.estimate_kld(),
         'mean-ess': problem.estimate_mean_ess()
     }
 
 
-def banana(n_iterations=5000, n_warmup=1000):
+def banana(n_iterations=4000, n_warmup=1000):
     """
-    Tests :class:`pints.DreamMCMC`
+    Tests :class:`pints.DramACMC`
     on a two-dimensional "twisted Gaussian" distribution with true solution
     ``[0, 0]`` and returns a dictionary with entries ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnBanana`.
+    :class:`pints.cptests.RunMcmcMethodOnBanana`.
     """
     problem = ft.RunMcmcMethodOnBanana(
-        _method, 20, n_iterations, n_warmup)
+        _method, 4, n_iterations, n_warmup)
     return {
         'kld': problem.estimate_kld(),
         'mean-ess': problem.estimate_mean_ess()
     }
 
 
-def correlated_gaussian(n_iterations=10000, n_warmup=1000):
+def correlated_gaussian(n_iterations=8000, n_warmup=4000):
     """
-    Tests :class:`pints.DreamMCMC`
+    Tests :class:`pints.DramACMC`
     on a six-dimensional highly correlated Gaussian distribution with true
     solution ``[0, 0, 0, 0, 0, 0]`` and returns a dictionary with entries
     ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnCorrelatedGaussian`.
+    :class:`pints.cptests.RunMcmcMethodOnCorrelatedGaussian`.
     """
     problem = ft.RunMcmcMethodOnCorrelatedGaussian(
-        _method, 20, n_iterations, n_warmup)
+        _method, 4, n_iterations, n_warmup)
     return {
         'kld': problem.estimate_kld(),
         'mean-ess': problem.estimate_mean_ess()
     }
 
 
-def annulus(n_iterations=10000, n_warmup=1000):
-    """
-    Tests :class:`pints.DreamMCMC`
-    on a two-dimensional annulus distribution with radius 10, and returns a
-    dictionary with entries ``distance`` and ``mean-ess``.
-
-    For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnAnnulus`.
-    """
-    problem = ft.RunMcmcMethodOnAnnulus(
-        _method, 10, n_iterations, n_warmup)
-    return {
-        'distance': problem.estimate_distance(),
-        'mean-ess': problem.estimate_mean_ess()
-    }
-
-
-_method = pints.DreamMCMC
-_functional_tests = [
-    annulus,
+_method = pints.DramACMC
+_change_point_tests = [
     banana,
     correlated_gaussian,
     two_dim_gaussian,
