+++ pints/cptests/haario_bardenet_acmc.py
@@ -1,23 +1,23 @@
 #!/usr/bin/env python3
 #
-# Functional tests for MetropolisRandomWalkMCMC
+# Change point tests for HaarioBardenetACMC
 #
 # This file is part of PINTS (https://github.com/pints-team/pints/) which is
 # released under the BSD 3-clause license. See accompanying LICENSE.md for
 # copyright notice and full license details.
 #
 import pints
-import pints.functionaltests as ft
+import pints.cptests as ft
 
 
 def two_dim_gaussian(n_iterations=4000, n_warmup=1000):
     """
-    Tests :class:`pints.MetropolisRandomWalkMCMC`
+    Tests :class:`pints.HaarioBardenetACMC`
     on a two-dimensional Gaussian distribution with true solution
     ``[0, 0]`` and returns a dictionary with entries ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnTwoDimGaussian`.
+    :class:`pints.cptests.RunMcmcMethodOnTwoDimGaussian`.
     """
     problem = ft.RunMcmcMethodOnTwoDimGaussian(
         _method, 4, n_iterations, n_warmup)
@@ -29,12 +29,12 @@ def two_dim_gaussian(n_iterations=4000, n_warmup=1000):
 
 def banana(n_iterations=4000, n_warmup=1000):
     """
-    Tests :class:`pints.MetropolisRandomWalkMCMC`
+    Tests :class:`pints.HaarioBardenetACMC`
     on a two-dimensional "twisted Gaussian" distribution with true solution
     ``[0, 0]`` and returns a dictionary with entries ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnBanana`.
+    :class:`pints.cptests.RunMcmcMethodOnBanana`.
     """
     problem = ft.RunMcmcMethodOnBanana(
         _method, 4, n_iterations, n_warmup)
@@ -46,13 +46,13 @@ def banana(n_iterations=4000, n_warmup=1000):
 
 def correlated_gaussian(n_iterations=8000, n_warmup=4000):
     """
-    Tests :class:`pints.MetropolisRandomWalkMCMC`
+    Tests :class:`pints.HaarioBardenetACMC`
     on a six-dimensional highly correlated Gaussian distribution with true
     solution ``[0, 0, 0, 0, 0, 0]`` and returns a dictionary with entries
     ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnCorrelatedGaussian`.
+    :class:`pints.cptests.RunMcmcMethodOnCorrelatedGaussian`.
     """
     problem = ft.RunMcmcMethodOnCorrelatedGaussian(
         _method, 4, n_iterations, n_warmup)
@@ -62,8 +62,26 @@ def correlated_gaussian(n_iterations=8000, n_warmup=4000):
     }
 
 
-_method = pints.MetropolisRandomWalkMCMC
-_functional_tests = [
+def annulus(n_iterations=4000, n_warmup=2000):
+    """
+    Tests :class:`pints.HaarioBardenetACMC`
+    on a two-dimensional annulus distribution with radius 10, and returns a
+    dictionary with entries ``distance`` and ``mean-ess``.
+
+    For details of the solved problem, see
+    :class:`pints.cptests.RunMcmcMethodOnAnnulus`.
+    """
+    problem = ft.RunMcmcMethodOnAnnulus(
+        _method, 4, n_iterations, n_warmup)
+    return {
+        'distance': problem.estimate_distance(),
+        'mean-ess': problem.estimate_mean_ess()
+    }
+
+
+_method = pints.HaarioBardenetACMC
+_change_point_tests = [
+    annulus,
     banana,
     correlated_gaussian,
     two_dim_gaussian,
