+++ pints/cptests/hamiltonian_mcmc.py
@@ -1,13 +1,13 @@
 #!/usr/bin/env python3
 #
-# Functional tests for HamiltonianMCMC
+# Change point tests for HamiltonianMCMC
 #
 # This file is part of PINTS (https://github.com/pints-team/pints/) which is
 # released under the BSD 3-clause license. See accompanying LICENSE.md for
 # copyright notice and full license details.
 #
 import pints
-import pints.functionaltests as ft
+import pints.cptests as ft
 
 
 def two_dim_gaussian(n_iterations=1000, n_warmup=200):
@@ -17,7 +17,7 @@ def two_dim_gaussian(n_iterations=1000, n_warmup=200):
     ``[0, 0]`` and returns a dictionary with entries ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnTwoDimGaussian`.
+    :class:`pints.cptests.RunMcmcMethodOnTwoDimGaussian`.
     """
     problem = ft.RunMcmcMethodOnTwoDimGaussian(
         method=_method,
@@ -39,7 +39,7 @@ def banana(n_iterations=2000, n_warmup=1000):
     ``[0, 0]`` and returns a dictionary with entries ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnBanana`.
+    :class:`pints.cptests.RunMcmcMethodOnBanana`.
     """
     problem = ft.RunMcmcMethodOnBanana(
         method=_method,
@@ -61,7 +61,7 @@ def high_dim_gaussian(n_iterations=4000, n_warmup=1000):
     returns a dictionary with entries ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnHighDimensionalGaussian`.
+    :class:`pints.cptests.RunMcmcMethodOnHighDimensionalGaussian`.
     """
     problem = ft.RunMcmcMethodOnHighDimensionalGaussian(
         method=_method,
@@ -77,7 +77,7 @@ def high_dim_gaussian(n_iterations=4000, n_warmup=1000):
 
 
 _method = pints.HamiltonianMCMC
-_functional_tests = [
+_change_point_tests = [
     banana,
     high_dim_gaussian,
     two_dim_gaussian,
