+++ pints/cptests/metropolis_random_walk_mcmc.py
@@ -1,23 +1,23 @@
 #!/usr/bin/env python3
 #
-# Functional tests for NoUTurnMCMC
+# Change point tests for MetropolisRandomWalkMCMC
 #
 # This file is part of PINTS (https://github.com/pints-team/pints/) which is
 # released under the BSD 3-clause license. See accompanying LICENSE.md for
 # copyright notice and full license details.
 #
 import pints
-import pints.functionaltests as ft
+import pints.cptests as ft
 
 
-def two_dim_gaussian(n_iterations=1000, n_warmup=200):
+def two_dim_gaussian(n_iterations=4000, n_warmup=1000):
     """
-    Tests :class:`pints.NoUTurnMCMC`
+    Tests :class:`pints.MetropolisRandomWalkMCMC`
     on a two-dimensional Gaussian distribution with true solution
     ``[0, 0]`` and returns a dictionary with entries ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnTwoDimGaussian`.
+    :class:`pints.cptests.RunMcmcMethodOnTwoDimGaussian`.
     """
     problem = ft.RunMcmcMethodOnTwoDimGaussian(
         _method, 4, n_iterations, n_warmup)
@@ -27,14 +27,14 @@ def two_dim_gaussian(n_iterations=1000, n_warmup=200):
     }
 
 
-def banana(n_iterations=2000, n_warmup=500):
+def banana(n_iterations=4000, n_warmup=1000):
     """
-    Tests :class:`pints.NoUTurnMCMC`
+    Tests :class:`pints.MetropolisRandomWalkMCMC`
     on a two-dimensional "twisted Gaussian" distribution with true solution
     ``[0, 0]`` and returns a dictionary with entries ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnBanana`.
+    :class:`pints.cptests.RunMcmcMethodOnBanana`.
     """
     problem = ft.RunMcmcMethodOnBanana(
         _method, 4, n_iterations, n_warmup)
@@ -44,16 +44,17 @@ def banana(n_iterations=2000, n_warmup=500):
     }
 
 
-def high_dim_gaussian(n_iterations=4000, n_warmup=1000):
+def correlated_gaussian(n_iterations=8000, n_warmup=4000):
     """
-     Tests :class:`pints.NoUTurnMCMC`
-    on a 20-dimensional Gaussian distribution centered at the origin, and
-    returns a dictionary with entries ``kld`` and ``mean-ess``.
+    Tests :class:`pints.MetropolisRandomWalkMCMC`
+    on a six-dimensional highly correlated Gaussian distribution with true
+    solution ``[0, 0, 0, 0, 0, 0]`` and returns a dictionary with entries
+    ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnHighDimensionalGaussian`.
+    :class:`pints.cptests.RunMcmcMethodOnCorrelatedGaussian`.
     """
-    problem = ft.RunMcmcMethodOnHighDimensionalGaussian(
+    problem = ft.RunMcmcMethodOnCorrelatedGaussian(
         _method, 4, n_iterations, n_warmup)
     return {
         'kld': problem.estimate_kld(),
@@ -61,9 +62,9 @@ def high_dim_gaussian(n_iterations=4000, n_warmup=1000):
     }
 
 
-_method = pints.NoUTurnMCMC
-_functional_tests = [
+_method = pints.MetropolisRandomWalkMCMC
+_change_point_tests = [
     banana,
-    high_dim_gaussian,
+    correlated_gaussian,
     two_dim_gaussian,
 ]
