+++ pints/cptests/monomial_gamma_hamiltonian_mcmc.py
@@ -1,30 +1,30 @@
 #!/usr/bin/env python3
 #
-# Functional tests for MALA MCMC
+# Change point tests for MonomialGammaHamiltonianMCMC
 #
 # This file is part of PINTS (https://github.com/pints-team/pints/) which is
 # released under the BSD 3-clause license. See accompanying LICENSE.md for
 # copyright notice and full license details.
 #
 import pints
-import pints.functionaltests as ft
+import pints.cptests as ft
 
 
 def two_dim_gaussian(n_iterations=1000, n_warmup=200):
     """
-    Tests :class:`pints.MALAMCMC`
+    Tests :class:`pints.MonomialGammaHamiltonianMCMC`
     on a two-dimensional Gaussian distribution with true solution
     ``[0, 0]`` and returns a dictionary with entries ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnTwoDimGaussian`.
+    :class:`pints.cptests.RunMcmcMethodOnTwoDimGaussian`.
     """
     problem = ft.RunMcmcMethodOnTwoDimGaussian(
         method=_method,
         n_chains=4,
         n_iterations=n_iterations,
         n_warmup=n_warmup,
-        method_hyper_parameters=[[1.0, 1.0]]
+        method_hyper_parameters=[20, 1, 0.5, 0.2, 1]
     )
     return {
         'kld': problem.estimate_kld(),
@@ -34,19 +34,19 @@ def two_dim_gaussian(n_iterations=1000, n_warmup=200):
 
 def banana(n_iterations=2000, n_warmup=500):
     """
-    Tests :class:`pints.MALAMCMC`
+    Tests :class:`pints.MonomialGammaHamiltonianMCMC`
     on a two-dimensional "twisted Gaussian" distribution with true solution
     ``[0, 0]`` and returns a dictionary with entries ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnBanana`.
+    :class:`pints.cptests.RunMcmcMethodOnBanana`.
     """
     problem = ft.RunMcmcMethodOnBanana(
         method=_method,
         n_chains=4,
         n_iterations=n_iterations,
         n_warmup=n_warmup,
-        method_hyper_parameters=[[0.8] * 2]
+        method_hyper_parameters=[20, 1, 0.5, 0.2, 1]
     )
     return {
         'kld': problem.estimate_kld(),
@@ -56,21 +56,20 @@ def banana(n_iterations=2000, n_warmup=500):
 
 def high_dim_gaussian(n_iterations=2000, n_warmup=500):
     """
-     Tests :class:`pints.MALAMCMC`
+     Tests :class:`pints.MonomialGammaHamiltonianMCMC`
     on a 20-dimensional Gaussian distribution centered at the origin, and
     returns a dictionary with entries ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnHighDimensionalGaussian`.
+    :class:`pints.cptests.RunMcmcMethodOnHighDimensionalGaussian`.
     """
     problem = ft.RunMcmcMethodOnHighDimensionalGaussian(
         method=_method,
         n_chains=4,
         n_iterations=n_iterations,
         n_warmup=n_warmup,
-        method_hyper_parameters=[[1.2] * 20]
+        method_hyper_parameters=[20, 1, 0.5, 0.2, 1]
     )
-
     return {
         'kld': problem.estimate_kld(),
         'mean-ess': problem.estimate_mean_ess()
@@ -79,20 +78,20 @@ def high_dim_gaussian(n_iterations=2000, n_warmup=500):
 
 def correlated_gaussian(n_iterations=2000, n_warmup=500):
     """
-    Tests :class:`pints.MALAMCMC`
+    Tests :class:`pints.MonomialGammaHamiltonianMCMC`
     on a six-dimensional highly correlated Gaussian distribution with true
     solution ``[0, 0, 0, 0, 0, 0]`` and returns a dictionary with entries
     ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnCorrelatedGaussian`.
+    :class:`pints.cptests.RunMcmcMethodOnCorrelatedGaussian`.
     """
     problem = ft.RunMcmcMethodOnCorrelatedGaussian(
         method=_method,
         n_chains=4,
         n_iterations=n_iterations,
         n_warmup=n_warmup,
-        method_hyper_parameters=[[1.0] * 6]
+        method_hyper_parameters=[20, 1, 0.5, 0.2, 1]
     )
     return {
         'kld': problem.estimate_kld(),
@@ -102,19 +101,19 @@ def correlated_gaussian(n_iterations=2000, n_warmup=500):
 
 def annulus(n_iterations=2000, n_warmup=500):
     """
-    Tests :class:`pints.MALAMCMC`
+    Tests :class:`pints.MonomialGammaHamiltonianMCMC`
     on a two-dimensional annulus distribution with radius 10, and returns a
     dictionary with entries ``distance`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnAnnulus`.
+    :class:`pints.cptests.RunMcmcMethodOnAnnulus`.
     """
     problem = ft.RunMcmcMethodOnAnnulus(
         method=_method,
         n_chains=4,
         n_iterations=n_iterations,
         n_warmup=n_warmup,
-        method_hyper_parameters=[[1.2] * 2],
+        method_hyper_parameters=[20, 1, 0.5, 0.2, 1]
     )
     return {
         'distance': problem.estimate_distance(),
@@ -124,20 +123,20 @@ def annulus(n_iterations=2000, n_warmup=500):
 
 def multimodal_gaussian(n_iterations=2000, n_warmup=500):
     """
-    Tests :class:`pints.MALAMCMC`
+    Tests :class:`pints.MonomialGammaHamiltonianMCMC`
     on a two-dimensional multi-modal Gaussian distribution with modes at
     ``[0, 0]``, ``[5, 10]``, and ``[10, 0]``, and returns a dict with entries
     "kld" and "mean-ess".
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnMultimodalGaussian`.
+    :class:`pints.cptests.RunMcmcMethodOnMultimodalGaussian`.
     """
     problem = ft.RunMcmcMethodOnMultimodalGaussian(
         method=_method,
         n_chains=4,
         n_iterations=n_iterations,
         n_warmup=n_warmup,
-        method_hyper_parameters=[[2.0] * 2]
+        method_hyper_parameters=[20, 1, 0.5, 0.2, 1]
     )
     return {
         'kld': problem.estimate_kld(),
@@ -147,19 +146,19 @@ def multimodal_gaussian(n_iterations=2000, n_warmup=500):
 
 def cone(n_iterations=2000, n_warmup=500):
     """
-    Tests :class:`pints.MALAMCMC`
+    Tests :class:`pints.MonomialGammaHamiltonianMCMC`
     on a two-dimensional cone distribution centered at ``[0, 0]``, and returns
     a dict with entries "distance" and "mean-ess".
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnCone`.
+    :class:`pints.cptests.RunMcmcMethodOnCone`.
     """
     problem = ft.RunMcmcMethodOnCone(
         method=_method,
         n_chains=4,
         n_iterations=n_iterations,
         n_warmup=n_warmup,
-        method_hyper_parameters=[[1.0, 1.0]],
+        method_hyper_parameters=[20, 1, 0.5, 0.2, 1]
     )
     return {
         'distance': problem.estimate_distance(),
@@ -167,8 +166,8 @@ def cone(n_iterations=2000, n_warmup=500):
     }
 
 
-_method = pints.MALAMCMC
-_functional_tests = [
+_method = pints.MonomialGammaHamiltonianMCMC
+_change_point_tests = [
     annulus,
     banana,
     cone,
