+++ pints/cptests/slice_stepout_mcmc.py
@@ -1,125 +1,143 @@
 #!/usr/bin/env python3
 #
-# Functional tests for EmceeHammerMCMC
+# Change point tests for SliceStepoutMCMC
 #
 # This file is part of PINTS (https://github.com/pints-team/pints/) which is
 # released under the BSD 3-clause license. See accompanying LICENSE.md for
 # copyright notice and full license details.
 #
 import pints
-import pints.functionaltests as ft
+import pints.cptests as ft
 
 
-def two_dim_gaussian(n_iterations=10000, n_warmup=1000):
+def two_dim_gaussian(n_iterations=5000, n_warmup=500):
     """
-    Tests :class:`pints.EmceeHammerMCMC`
+    Tests :class:`pints.SliceStepoutMCMC`
     on a two-dimensional Gaussian distribution with true solution
     ``[0, 0]`` and returns a dictionary with entries ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnTwoDimGaussian`.
+    :class:`pints.cptests.RunMcmcMethodOnTwoDimGaussian`.
     """
     problem = ft.RunMcmcMethodOnTwoDimGaussian(
-        _method, 10, n_iterations, n_warmup)
+        _method, 4, n_iterations, n_warmup)
     return {
         'kld': problem.estimate_kld(),
         'mean-ess': problem.estimate_mean_ess()
     }
 
 
-def banana(n_iterations=10000, n_warmup=2000):
+def correlated_gaussian(n_iterations=5000, n_warmup=500):
     """
-    Tests :class:`pints.EmceeHammerMCMC`
-    on a two-dimensional "twisted Gaussian" distribution with true solution
-    ``[0, 0]`` and returns a dictionary with entries ``kld`` and ``mean-ess``.
-
-    For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnBanana`.
-    """
-    problem = ft.RunMcmcMethodOnBanana(
-        _method, 10, n_iterations, n_warmup)
-    return {
-        'kld': problem.estimate_kld(),
-        'mean-ess': problem.estimate_mean_ess()
-    }
-
-
-def correlated_gaussian(n_iterations=8000, n_warmup=4000):
-    """
-    Tests :class:`pints.EmceeHammerMCMC`
+    Tests :class:`pints.SliceStepoutMCMC`
     on a six-dimensional highly correlated Gaussian distribution with true
     solution ``[0, 0, 0, 0, 0, 0]`` and returns a dictionary with entries
     ``kld`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnCorrelatedGaussian`.
+    :class:`pints.cptests.RunMcmcMethodOnCorrelatedGaussian`.
     """
     problem = ft.RunMcmcMethodOnCorrelatedGaussian(
-        _method, 10, n_iterations, n_warmup)
+        _method, 4, n_iterations, n_warmup)
     return {
         'kld': problem.estimate_kld(),
         'mean-ess': problem.estimate_mean_ess()
     }
 
 
-def annulus(n_iterations=4000, n_warmup=2000):
+def banana(n_iterations=5000, n_warmup=500):
     """
-    Tests :class:`pints.EmceeHammerMCMC`
+    Tests :class:`pints.SliceStepoutMCMC`
+    on a two-dimensional "twisted Gaussian" distribution with true solution
+    ``[0, 0]`` and returns a dictionary with entries ``kld`` and ``mean-ess``.
+
+    For details of the solved problem, see
+    :class:`pints.cptests.RunMcmcMethodOnBanana`.
+    """
+    problem = ft.RunMcmcMethodOnBanana(
+        _method, 4, n_iterations, n_warmup)
+    return {
+        'kld': problem.estimate_kld(),
+        'mean-ess': problem.estimate_mean_ess()
+    }
+
+
+def high_dim_gaussian(n_iterations=5000, n_warmup=500):
+    """
+     Tests :class:`pints.SliceStepoutMCMC`
+    on a 20-dimensional Gaussian distribution centered at the origin, and
+    returns a dictionary with entries ``kld`` and ``mean-ess``.
+
+    For details of the solved problem, see
+    :class:`pints.cptests.RunMcmcMethodOnHighDimensionalGaussian`.
+    """
+    problem = ft.RunMcmcMethodOnHighDimensionalGaussian(
+        _method, 4, n_iterations, n_warmup)
+    return {
+        'kld': problem.estimate_kld(),
+        'mean-ess': problem.estimate_mean_ess()
+    }
+
+
+def annulus(n_iterations=10000, n_warmup=2000):
+    """
+    Tests :class:`pints.SliceStepoutMCMC`
     on a two-dimensional annulus distribution with radius 10, and returns a
     dictionary with entries ``distance`` and ``mean-ess``.
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnAnnulus`.
+    :class:`pints.cptests.RunMcmcMethodOnAnnulus`.
     """
     problem = ft.RunMcmcMethodOnAnnulus(
-        _method, 10, n_iterations, n_warmup)
+        _method, 4, n_iterations, n_warmup)
     return {
         'distance': problem.estimate_distance(),
         'mean-ess': problem.estimate_mean_ess()
     }
 
 
-def multimodal_gaussian(n_iterations=10000, n_warmup=1000):
+def multimodal_gaussian(n_iterations=5000, n_warmup=500):
     """
-    Tests :class:`pints.EmceeHammerMCMC`
+    Tests :class:`pints.SliceStepoutMCMC`
     on a two-dimensional multi-modal Gaussian distribution with modes at
     ``[0, 0]``, ``[5, 10]``, and ``[10, 0]``, and returns a dict with entries
     "kld" and "mean-ess".
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnMultimodalGaussian`.
+    :class:`pints.cptests.RunMcmcMethodOnMultimodalGaussian`.
     """
     problem = ft.RunMcmcMethodOnMultimodalGaussian(
-        _method, 10, n_iterations, n_warmup)
+        _method, 4, n_iterations, n_warmup)
     return {
         'kld': problem.estimate_kld(),
         'mean-ess': problem.estimate_mean_ess()
     }
 
 
-def cone(n_iterations=10000, n_warmup=1000):
+def cone(n_iterations=5000, n_warmup=500):
     """
-    Tests :class:`pints.EmceeHammerMCMC`
+    Tests :class:`pints.SliceStepoutMCMC`
     on a two-dimensional cone distribution centered at ``[0, 0]``, and returns
     a dict with entries "distance" and "mean-ess".
 
     For details of the solved problem, see
-    :class:`pints.functionaltests.RunMcmcMethodOnCone`.
+    :class:`pints.cptests.RunMcmcMethodOnCone`.
     """
     problem = ft.RunMcmcMethodOnCone(
-        _method, 10, n_iterations, n_warmup)
+        _method, 4, n_iterations, n_warmup)
     return {
         'distance': problem.estimate_distance(),
         'mean-ess': problem.estimate_mean_ess()
     }
 
 
-_method = pints.EmceeHammerMCMC
-_functional_tests = [
+_method = pints.SliceStepoutMCMC
+_change_point_tests = [
     annulus,
     banana,
     cone,
     correlated_gaussian,
+    high_dim_gaussian,
     multimodal_gaussian,
     two_dim_gaussian,
 ]
